backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
openai==1.40.0